# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/status:
    get:
      tags: [Auth]
      operationId: getAdminStatus
      summary: Whether the current user is the app owner (admin)
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Admin status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStatusEnvelope'
  /admin/users:
    get:
      tags: [Admin]
      operationId: listAdminUsers
      summary: List all users with a fitness summary (admin only)
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: The list of users.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserListEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Not an admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/users/{userId}:
    get:
      tags: [Admin]
      operationId: getAdminUserDetail
      summary: Full detail for one user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: The user detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserDetailEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Not an admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Admin]
      operationId: updateAdminUser
      summary: Update a user's credits or goal (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAdminUserRequest'
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Updated user detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserDetailEnvelope'
        '400':
          description: Invalid request, or the user has no saved data to edit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Not an admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/users/{userId}/remind:
    post:
      tags: [Admin]
      operationId: sendUserReminder
      summary: Email a user a nudge to finish onboarding (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: The reminder email was sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendReminderResult'
        '400':
          description: The user has no email address on file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Not an admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: The email could not be sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    AdminStatusEnvelope:
      type: object
      required: [isAdmin, isOwner]
      properties:
        isAdmin:
          type: boolean
        isOwner:
          type: boolean
    AdminCredits:
      type: object
      required: [coaching, photo, bodyScan]
      properties:
        coaching:
          type: integer
        photo:
          type: integer
        bodyScan:
          type: integer
    AdminUserSummary:
      type: object
      required:
        - id
        - email
        - username
        - name
        - goal
        - onboardingComplete
        - planDays
        - credits
        - createdAt
        - lastActive
        - photoCount
        - prCount
        - weightLogCount
        - latestWeight
        - bodyFat
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        username:
          type: [string, 'null']
        name:
          type: string
        goal:
          type: [string, 'null']
        onboardingComplete:
          type: boolean
        planDays:
          type: integer
        credits:
          oneOf:
            - $ref: '#/components/schemas/AdminCredits'
            - type: 'null'
        createdAt:
          type: string
          format: date-time
        lastActive:
          type: [string, 'null']
          format: date-time
        photoCount:
          type: integer
        prCount:
          type: integer
        weightLogCount:
          type: integer
        latestWeight:
          type: [number, 'null']
        bodyFat:
          type: [number, 'null']
    AdminUserListEnvelope:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/AdminUserSummary'
    AdminUserDetail:
      type: object
      required:
        - summary
        - profile
        - plan
        - programMeta
        - prs
        - weightLogs
        - progressPhotos
      properties:
        summary:
          $ref: '#/components/schemas/AdminUserSummary'
        profile:
          oneOf:
            - type: object
              additionalProperties: true
            - type: 'null'
        plan:
          type: array
          items:
            type: object
            additionalProperties: true
        programMeta:
          oneOf:
            - type: object
              additionalProperties: true
            - type: 'null'
        prs:
          type: array
          items:
            type: object
            additionalProperties: true
        weightLogs:
          type: array
          items:
            type: object
            additionalProperties: true
        progressPhotos:
          type: array
          items:
            type: object
            additionalProperties: true
    AdminUserDetailEnvelope:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/AdminUserDetail'
    UpdateAdminUserRequest:
      type: object
      properties:
        credits:
          $ref: '#/components/schemas/AdminCredits'
        goal:
          type: [string, 'null']
          enum: ['Weight Loss', 'Muscle Gain', 'Strength', 'Athleticism', null]
    SendReminderResult:
      type: object
      required: [success, sentTo]
      properties:
        success:
          type: boolean
        sentTo:
          type: string
          description: The email address the reminder was sent to.
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string